raulillo lo a coseguido

+++ nautico/src/main/java/club/nautico/persistence/repository/InterfazSalida.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import club.nautico.persistence.entity.Salida;
 
-public interface InterfazSalida extends CrudRepository<Salida, Long >{
+public interface InterfazSalida extends CrudRepository<Salida, Integer >{
 
 }
